+++ cumulus/client/consensus/aura/src/collators/slot_based/block_import.rs
@@ -18,7 +18,8 @@
 use crate::LOG_TARGET;
 use codec::{Decode, Encode};
 use cumulus_client_proof_size_recording::prepare_proof_size_recording_aux_data;
-use cumulus_client_unincluded_segment_store::{now_unix_ms, old_finalized_hash, prepare_unincluded_segment_aux_data};
+use cumulus_client_consensus_common::old_finalized_hash;
+use cumulus_client_unincluded_segment_store::{now_unix_ms, prepare_unincluded_segment_aux_data};
 use cumulus_primitives_core::{BlockBundleInfo, CoreInfo, CumulusDigestItem, RelayBlockIdentifier};
 use futures::{stream::FusedStream, StreamExt};
 use sc_client_api::{
