+++ cumulus/client/consensus/common/src/lib.rs
@@ -30,12 +30,14 @@ use sp_timestamp::Timestamp;
 
 use std::{sync::Arc, time::Duration};
 
+mod finality;
 mod level_monitor;
 mod parachain_consensus;
 mod parent_search;
 #[cfg(test)]
 mod tests;
 
+pub use finality::old_finalized_hash;
 pub use parent_search::*;
 
 pub use cumulus_relay_chain_streams::finalized_heads;
