+++ cumulus/client/unincluded-segment-store/Cargo.toml
@@ -13,6 +13,7 @@ workspace = true
 
 [dependencies]
 codec = { workspace = true, default-features = true }
+cumulus-client-consensus-common = { workspace = true, default-features = true }
 sc-client-api = { workspace = true, default-features = true }
 sp-blockchain = { workspace = true, default-features = true }
 sp-runtime = { workspace = true, default-features = true }
